Create Exercise Controller & Service. Add Get & GetByID methods.

+++ src/main/java/be/thomasmol/workout_tracker/exercise/ExerciseRepository.java
@@ -2,5 +2,5 @@ package be.thomasmol.workout_tracker.exercise;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExerciseRepository extends JpaRepository<Exercise, Integer> {
+public interface ExerciseRepository extends JpaRepository<Exercise, Long> {
 }
